refactor: move types to the unlock module
